Crates clientlib and configuration merged into bot-core
[GH-34]
