+++ bot-core/src/configuration.rs
