+++ bot-core/src/users/client_meta.rs
