+++ bot-core/src/users/subscriptions.rs
