--- clientlib/.env
@@ -1,1 +1,0 @@
-export DATABASE_URL=mariadb://root:password@localhost:3306/shortbot
