--- clientlib/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "clientlib"
-edition.workspace = true
-version.workspace = true
-license.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-tracing = { version = "0.1", features = ["log"] }
-sqlx = { version = "0.8.6", features = [ "runtime-tokio", "mysql", "macros", "chrono" ] }
-secrecy = { version = "0.10.3", features = ["serde"] }
-uuid = { version = "1.13.1", features = ["v4"] }
-tokio = {version = "1.8", features = ["rt", "macros", "sync"]}
-teloxide = { version = "0.16.0" }
-thiserror = "2.0.12"
-whirlwind = "0.1.1"
-chrono = { version = "0.4.38", features = ["alloc", "std", "clock"] }
-
-[dev-dependencies]
-configuration = { path = "../configuration" }
-random = "0.14.0"
-tracing-subscriber = "0.3.19"
-once_cell = "1.21.3"
-
-[lints]
-workspace = true
